函数移位
